+++ src/renderer/features/wallets/KeyConstructor/index.ts
@@ -1,3 +1,2 @@
 export { KeyConstructor } from './ui/KeyConstructor';
 export { type DerivationKeyDraft } from './model/constructor-model';
-export { populateDraftAccounts } from './lib/utils';
